+++ src/index.js
@@ -8,6 +8,7 @@ import {
   styleSearch,
   validateOptions,
 } from "./utils"
+import ruleTester from "./testUtils/ruleTester"
 
 const stylelint = postcssPlugin
 
@@ -16,6 +17,7 @@ stylelint.utils = {
   ruleMessages,
   styleSearch,
   validateOptions,
+  ruleTester,
 }
 
 stylelint.lint = standalone